requirements.txt
# Core Data Science
numpy>=1.21.0
pandas>=1.3.0
scikit-learn>=1.0.0
xgboost>=1.5.0
imbalanced-learn>=0.8.0

# Data Processing
dask>=2021.10.0
pyarrow>=6.0.0
joblib>=1.1.0
python-dateutil>=2.8.2
pytz>=2021.3

# Stream Processing
kafka-python>=2.0.2
pyspark>=3.2.0

# Visualization
matplotlib>=3.4.0
seaborn>=0.11.0
plotly>=5.3.0
dash>=2.0.0
dash-bootstrap-components>=1.0.0

# Utilities
tqdm>=4.62.0
python-dotenv>=0.19.0
pyyaml>=6.0

# Development
jupyter>=1.0.0
pytest>=6.2.5
black>=21.12b0
flake8>=4.0.0

# Web/Dashboard
flask>=2.0.0
gunicorn>=20.1.0

# Optional: For geo-spatial features
geopy>=2.2.0
haversine>=2.5.0
